% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{add}
\alias{add}
\title{Sum}
\usage{
add(x, y, plot = FALSE)
}
\arguments{
\item{x, y}{\code{numeric} vectors.}

\item{plot}{\code{FALSE} so no plot was used.}
}
\value{
\code{numeric} vector of \code{x}, \code{y}and \code{x} +\code{y}.
}
\description{
Sum
}
\examples{
#sum of two variable
add(1,4)
# sum of two vectors
add(c(1:2),1:8)

}
