% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pattern_to_ms}
\alias{pattern_to_ms}
\title{Convert a given rhythm pattern, a BPM and a BPM type to duration values}
\usage{
pattern_to_ms(pattern, bpm = 120, type)
}
\arguments{
\item{type}{}
}
\description{
Convert a given rhythm pattern, a BPM and a BPM type to duration values
}
