% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{rhythm_feedback}
\alias{rhythm_feedback}
\title{Setup feedback}
\usage{
rhythm_feedback(
  type,
  text = "Well done!",
  image = "https://adaptiveeartraining.com/assets/drum.png",
  height = NULL,
  width = NULL,
  progress = NULL,
  get_async_data = FALSE
)
}
\arguments{
\item{get_async_data}{}
}
\description{
Setup feedback
}
