% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTT.R
\name{RTT_standalone}
\alias{RTT_standalone}
\title{Launch the Rhythm Tapping Task}
\usage{
RTT_standalone(
  app_name = "RTT",
  musicassessr_aws = FALSE,
  data_collection_method = c("midi", "audio", "key_presses")
)
}
\arguments{
\item{data_collection_method}{}
}
\description{
Launch the Rhythm Tapping Task
}
