% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTT.R
\name{RTT}
\alias{RTT}
\title{Rhythm tapping test}
\usage{
RTT(
  page_type = "record_midi_page",
  feedback = rhythm_feedback(type = "none"),
  num_items = list(free_recall = 3L, sync_beat = 3L, call_and_response = 3L),
  num_examples = list(free_recall = 2L, sync_beat = 2L, call_and_response = 2L),
  call_and_response_end = c("manual", "auto"),
  call_and_response_bpm = "user_determined",
  setup_pages = TRUE,
  data_collection_method = c("midi", "audio", "key_presses"),
  get_p_id = TRUE
)
}
\arguments{
\item{get_p_id}{}
}
\description{
Rhythm tapping test
}
