% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTT.R
\name{RTT}
\alias{RTT}
\title{Rhythm tapping test}
\usage{
RTT(
  page_type = "record_midi_page",
  feedback = rhythm_feedback(type = "none"),
  num_items = list(free_recall = 3L, sync_beat = 3L, call_and_response = 3L),
  num_examples = list(free_recall = 2L, sync_beat = 2L, call_and_response = 2L),
  call_and_response_end = c("manual", "auto"),
  call_and_response_bpm = "user_determined",
  setup_pages = TRUE,
  data_collection_method = c("midi", "audio", "key_presses"),
  get_p_id = TRUE,
  sync_beat_bpm_range = 60:200,
  sync_beat_bpms = NULL,
  filter_call_and_response_stimuli_length = NULL,
  mute_midi_playback = TRUE,
  standalone = FALSE,
  asynchronous_api_mode = FALSE,
  app_name,
  experiment_id = NULL,
  user_id = NULL,
  feedback_free_recall = rhythm_feedback(type = "none")
)
}
\arguments{
\item{sync_beat_bpms}{A vector of BPMs.}

\item{filter_call_and_response_stimuli_length}{NULL = not filtered. Or a vector of the range of stimulus lengths. e.g., 3:12 will only have stimulus lengths 3-12.}

\item{mute_midi_playback}{Should user audio feedback be muted when using a MIDI pad?}

\item{feedback_free_recall}{}
}
\description{
Rhythm tapping test
}
