% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_battery.R
\name{RTT_calibration_constructor}
\alias{RTT_calibration_constructor}
\title{Constructor for 2023 calibration study}
\usage{
RTT_calibration_constructor(
  app_name = "RTT_calibration_study",
  data_collection_method = "audio",
  musicassessr_aws = FALSE
)
}
\arguments{
\item{musicassessr_aws}{}
}
\description{
Constructor for 2023 calibration study
}
