% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTT.R
\name{RTT_standalone}
\alias{RTT_standalone}
\title{Launch the Rhythm Tapping Task}
\usage{
RTT_standalone(
  app_name = "RTT",
  musicassessr_aws = FALSE,
  data_collection_method = c("midi", "audio", "key_presses"),
  feedback = rhythm_feedback(type = "none"),
  num_items = list(free_recall = 3L, sync_beat = 3L, call_and_response = 10L),
  num_examples = list(free_recall = 2L, sync_beat = 2L, call_and_response = 2L),
  call_and_response_end = c("manual", "auto"),
  call_and_response_bpm = "user_determined",
  get_p_id = TRUE
)
}
\arguments{
\item{get_p_id}{}
}
\description{
Launch the Rhythm Tapping Task
}
