% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTT.R
\name{RTT_standalone}
\alias{RTT_standalone}
\title{Launch the Rhythm Tapping Task}
\usage{
RTT_standalone(
  app_name = "RTT",
  data_collection_method = c("midi", "audio", "key_presses"),
  feedback = rhythm_feedback(type = "none"),
  num_items = list(free_recall = 3L, sync_beat = 3L, call_and_response = 10L),
  num_examples = list(free_recall = 2L, sync_beat = 2L, call_and_response = 2L),
  call_and_response_end = c("manual", "auto"),
  call_and_response_bpm = "user_determined",
  get_p_id = TRUE,
  sync_beat_bpm_range = 60:200,
  sync_beat_bpms = NULL,
  filter_call_and_response_stimuli_length = NULL,
  opening_and_final_image = "https://adaptiveeartraining.com/assets/drum.png",
  mute_midi_playback = TRUE,
  asynchronous_api_mode = FALSE,
  experiment_id = NULL,
  user_id = NULL,
  use_presigned_url = FALSE,
  feedback_free_recall = rhythm_feedback(type = "none"),
  SNR_test = FALSE,
  allow_SNR_failure = FALSE
)
}
\arguments{
\item{sync_beat_bpms}{A vector of BPMs.}

\item{filter_call_and_response_stimuli_length}{NULL = not filtered. Or a vector of the range of stimulus lengths. e.g., 3:12 will only have stimulus lengths 3-12.}

\item{opening_and_final_image}{A string to the location of the image to use on the first and final pages.}

\item{mute_midi_playback}{Should user audio feedback be muted when using a MIDI pad?}

\item{allow_SNR_failure}{}
}
\description{
Launch the Rhythm Tapping Task
}
